+++ utils.c
@@ -8,7 +8,7 @@
 #include <stdio.h>
 #include <string.h>
 
-#include "TinySmartAutomation.h"
+#include "utils.h"
 
 
 /**
